Create AnimationScreen abstract class
Helps reduce the size and complexity of FlashbackPlayer.

On a different note, it seems the odd animation reversal happens only
after you see a flashback for the first time, which is very odd.

Wait a second, oh my god, I think I figured out what is happening!

The player is literally seeing a memory of the flashback, that is so
funny! How did I not notice this until I started typing this out! At
least I know now how to fix it!