Add mask animation (#2)
* 🔥 Add basic mask animation

For an unknown reason, despite debugging saying otherwise, the
animation sometimes suddenly goes backwards right before the end!

* Create AnimationScreen abstract class

Helps reduce the size and complexity of FlashbackPlayer.

On a different note, it seems the odd animation reversal happens only
after you see a flashback for the first time, which is very odd.

Wait a second, oh my god, I think I figured out what is happening!

The player is literally seeing a memory of the flashback, that is so
funny! How did I not notice this until I started typing this out! At
least I know now how to fix it!

* Do not snap during flashback

It was causing to animation to appear to play backwards.

* Fix mask "zoom" animation

* Stop using the term "zoom" and adjust timings slightly

* Add fade in

I am nearing hour 22 according to IntelliJ time tracking!

* Use perspective instead of linear for memories

Also added a slight fade to them so they appear in the eye.

* Presumably read enchant particle sprites list

* Encapsulate the mask and memories in their own class

The FramedAnimationComponent class should help with creating timelined
components without worrying about when they actually start on a larger
timeline.

This should help with making the memory stream particles which will use
enchantment table glyphs.

* Create marginally acceptable glyph trails

They still need a lot of work on their timing and placement

* Attempt to make glyph trails ticker

* Work on spread and speed of glyph trails

* Safe render glyph trails and slow them down

Taking a more "personal" look for their existence

* Annotate and begin tweaking performance

* Hide the cursor on animation screens